@@ -4,27 +4,24 @@ publish_to: 'none'
 version: 0.1.0
 
 environment:
-  sdk: '>=3.0.5 <4.0.0'
+  sdk: '>=3.4.3 < 4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^1.1.0
+  http: ^1.2.1
   flutter_pagination: ^0.0.8
-  shared_preferences: ^2.2.2
-  google_fonts: ^5.1.0
+  shared_preferences: ^2.2.3
   provider: ^6.1.2
-  image_picker: ^1.0.7
-  flutter_pw_validator: ^1.5.0
-  intl: ^0.17.0
+  image_picker: ^1.1.2
+  flutter_pw_validator: ^1.6.0
+  intl: ^0.19.0
   dart_amqp: ^0.3.1
-  fluttertoast: ^8.0.2
-
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
 
 flutter:
   uses-material-design: true
