@@ -15,15 +15,13 @@ class FeedbackService {
     var token = await UserPreferences().getToken();
 
 
-    if (searchText != null && searchText.isNotEmpty) {
+    if (searchText.isNotEmpty) {
       uri += '&text=$searchText';
     }
 
     final response = await http.get(Uri.parse(uri),  headers: {
         'Authorization': 'Bearer $token',
       },);
-
-      print(response.statusCode);
 
     if (response.statusCode == 200) {
 
